hotfix : 특정 유저의 작성 게시글 조회 요청 형식 변경

+++ src/main/java/land/leets/Carrot/domain/post/repository/PostRepository.java
@@ -15,5 +15,5 @@ public interface PostRepository extends JpaRepository<Post, Long> {
     Optional<List<Post>> findByStatus(@Param("status") String status);
 
     @Query("SELECT post FROM Post post WHERE post.ceo.id=:userId")
-    Optional<List<Post>> findByWriterId(@Param("userId") Long userId);
+    Optional<List<Post>> findByWriterId(@Param("userId") Integer userId);
 }
